// FE/src/api/models/locationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * SIGREF API
 * API de Auditoría para Sistema de Gestión de Referencias FHIR
 * OpenAPI spec version: 1.0.0
 */

export type LocationStatus =
  (typeof LocationStatus)[keyof typeof LocationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LocationStatus = {
  active: "active",
  suspended: "suspended",
  inactive: "inactive",
} as const;